+++ server/src/controllers/users/postUsersDb.js
@@ -1,4 +1,4 @@
-const { User } = require("../db");
+const { User } = require("../../db");
 
 const postUser = async (name, surName, email, password) => {
     const maxId = await User.max('id');
